+++ node/src/GlideClient.ts
@@ -60,6 +60,7 @@ import {
     createScriptExists,
     createScriptFlush,
     createScriptKill,
+    createScriptDebug,
     createFailover,
     createReplicaOf,
     createReplicaOfNoOne,
@@ -77,6 +78,7 @@ import {
     LolwutOptions,
     MigrateOptions,
     ScanOptions,
+    ScriptDebugMode,
 } from "./Commands";
 
 /* eslint-disable-next-line @typescript-eslint/no-namespace */
@@ -1311,6 +1313,26 @@ export class GlideClient extends BaseClient {
      */
     public async scriptKill(): Promise<"OK"> {
         return this.createWritePromise(createScriptKill(), {
+            decoder: Decoder.String,
+        });
+    }
+
+    /**
+     * Set the debugging mode for executed scripts.
+     *
+     * @see {@link https://valkey.io/commands/script-debug/|valkey.io} for more details.
+     *
+     * @param mode - The debugging mode to set.
+     * @returns A simple `"OK"` response.
+     *
+     * @example
+     * ```typescript
+     * const result = await client.scriptDebug(ScriptDebugMode.YES);
+     * console.log(result); // Output: "OK"
+     * ```
+     */
+    public async scriptDebug(mode: ScriptDebugMode): Promise<"OK"> {
+        return this.createWritePromise(createScriptDebug(mode), {
             decoder: Decoder.String,
         });
     }